minor: move oauth access functions from email_account
